requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
bs4==0.0.1
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
lxml==4.9.2
mypy-extensions==1.0.0
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
python-dotenv==1.0.0
PySocks==1.7.1
requests==2.32.3
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
webdriver-manager==4.0.2
wsproto==1.2.0


pypdf
fuzzywuzzy
python-Levenshtein
reportlab
pytesseract
Pillow
pdfplumber>=0.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0
